+++ 05-JWT-Basics/starter/routes/main.js
@@ -1,9 +1,10 @@
 const express = require('express')
 const { login, dashboard } = require('../controllers/main')
+const authenticationMiddleware = require('../middleware/auth')
 
 const router = express.Router()
 
-router.route('/dashboard').get(dashboard)
+router.route('/dashboard').get(authenticationMiddleware, dashboard)
 router.route('/login').post(login)
 
 module.exports = router
